fix execute twice
