+++ (2022-06-16T04_48_20Z)_(2022-06-16T04_57_43Z).json
@@ -1,0 +1,302 @@
+[
+  {
+    "id": "2",
+    "time": [
+      2022,
+      6,
+      16,
+      4,
+      48,
+      20,
+      3,
+      167,
+      0
+    ],
+    "current": [
+      "1.1",
+      "1.2",
+      "1.1"
+    ],
+    "temperature": [
+      "23",
+      "24",
+      "23"
+    ],
+    "vibration": "45"
+  },
+  {
+    "id": "2",
+    "time": [
+      2022,
+      6,
+      16,
+      4,
+      50,
+      11,
+      3,
+      167,
+      0
+    ],
+    "current": [
+      "1.1",
+      "1.2",
+      "1.1"
+    ],
+    "temperature": [
+      "23",
+      "24",
+      "23"
+    ],
+    "vibration": "45"
+  },
+  {
+    "id": "2",
+    "time": [
+      2022,
+      6,
+      16,
+      4,
+      50,
+      11,
+      3,
+      167,
+      0
+    ],
+    "current": [
+      "1.1",
+      "1.2",
+      "1.1"
+    ],
+    "temperature": [
+      "23",
+      "24",
+      "23"
+    ],
+    "vibration": "45"
+  },
+  {
+    "id": "2",
+    "time": [
+      2022,
+      6,
+      16,
+      4,
+      51,
+      30,
+      3,
+      167,
+      0
+    ],
+    "current": [
+      "1.1",
+      "1.2",
+      "1.1"
+    ],
+    "temperature": [
+      "23",
+      "24",
+      "23"
+    ],
+    "vibration": "45"
+  },
+  {
+    "id": "2",
+    "time": [
+      2022,
+      6,
+      16,
+      4,
+      51,
+      35,
+      3,
+      167,
+      0
+    ],
+    "current": [
+      "1.1",
+      "1.2",
+      "1.1"
+    ],
+    "temperature": [
+      "23",
+      "24",
+      "23"
+    ],
+    "vibration": "45"
+  },
+  {
+    "id": "2",
+    "time": [
+      2022,
+      6,
+      16,
+      4,
+      51,
+      54,
+      3,
+      167,
+      0
+    ],
+    "current": [
+      "1.1",
+      "1.2",
+      "1.1"
+    ],
+    "temperature": [
+      "23",
+      "24",
+      "23"
+    ],
+    "vibration": "45"
+  },
+  {
+    "id": "2",
+    "time": [
+      2022,
+      6,
+      16,
+      4,
+      52,
+      44,
+      3,
+      167,
+      0
+    ],
+    "current": [
+      "1.1",
+      "1.2",
+      "1.1"
+    ],
+    "temperature": [
+      "23",
+      "24",
+      "23"
+    ],
+    "vibration": "45"
+  },
+  {
+    "id": "2",
+    "time": [
+      2022,
+      6,
+      16,
+      4,
+      52,
+      47,
+      3,
+      167,
+      0
+    ],
+    "current": [
+      "1.1",
+      "1.2",
+      "1.1"
+    ],
+    "temperature": [
+      "23",
+      "24",
+      "23"
+    ],
+    "vibration": "45"
+  },
+  {
+    "id": "2",
+    "time": [
+      2022,
+      6,
+      16,
+      4,
+      53,
+      4,
+      3,
+      167,
+      0
+    ],
+    "current": [
+      "1.1",
+      "1.2",
+      "1.1"
+    ],
+    "temperature": [
+      "23",
+      "24",
+      "23"
+    ],
+    "vibration": "45"
+  },
+  {
+    "id": "2",
+    "time": [
+      2022,
+      6,
+      16,
+      4,
+      56,
+      54,
+      3,
+      167,
+      0
+    ],
+    "current": [
+      "1.1",
+      "1.2",
+      "1.1"
+    ],
+    "temperature": [
+      "23",
+      "24",
+      "23"
+    ],
+    "vibration": "45"
+  },
+  {
+    "id": "2",
+    "time": [
+      2022,
+      6,
+      16,
+      4,
+      57,
+      42,
+      3,
+      167,
+      0
+    ],
+    "current": [
+      "1.1",
+      "1.2",
+      "1.1"
+    ],
+    "temperature": [
+      "23",
+      "24",
+      "23"
+    ],
+    "vibration": "45"
+  },
+  {
+    "id": "2",
+    "time": [
+      2022,
+      6,
+      16,
+      4,
+      57,
+      43,
+      3,
+      167,
+      0
+    ],
+    "current": [
+      "1.1",
+      "1.2",
+      "1.1"
+    ],
+    "temperature": [
+      "23",
+      "24",
+      "23"
+    ],
+    "vibration": "45"
+  }
+]
