+++ (2022-06-16T05_06_57Z)_(2022-06-16T05_06_57Z).json
@@ -1,0 +1,52 @@
+[
+  {
+    "id": "2",
+    "time": [
+      2022,
+      6,
+      16,
+      5,
+      6,
+      57,
+      3,
+      167,
+      0
+    ],
+    "current": [
+      "1.1",
+      "1.2",
+      "1.1"
+    ],
+    "temperature": [
+      "23",
+      "24",
+      "23"
+    ],
+    "vibration": "42"
+  },
+  {
+    "id": "2",
+    "time": [
+      2022,
+      6,
+      16,
+      5,
+      6,
+      57,
+      3,
+      167,
+      0
+    ],
+    "current": [
+      "1.1",
+      "1.2",
+      "1.1"
+    ],
+    "temperature": [
+      "23",
+      "24",
+      "23"
+    ],
+    "vibration": "42"
+  }
+]
