+++ (2022-06-16T05_07_15Z)_(2022-06-16T05_07_16Z).json
@@ -1,0 +1,52 @@
+[
+  {
+    "id": "2",
+    "time": [
+      2022,
+      6,
+      16,
+      5,
+      7,
+      15,
+      3,
+      167,
+      0
+    ],
+    "current": [
+      "1.1",
+      "1.2",
+      "1.1"
+    ],
+    "temperature": [
+      "23",
+      "24",
+      "23"
+    ],
+    "vibration": "42"
+  },
+  {
+    "id": "2",
+    "time": [
+      2022,
+      6,
+      16,
+      5,
+      7,
+      16,
+      3,
+      167,
+      0
+    ],
+    "current": [
+      "1.1",
+      "1.2",
+      "1.1"
+    ],
+    "temperature": [
+      "23",
+      "24",
+      "23"
+    ],
+    "vibration": "42"
+  }
+]
